更新sdk，更新闹钟、世界时钟等接口

+++ TopStepComKit-Git/Classes/Foundation/TopStepInterfaceKit.xcframework/ios-arm64/TopStepInterfaceKit.framework/Headers/TSSportItemModel.h
@@ -125,14 +125,14 @@ NS_ASSUME_NONNULL_BEGIN
  * @chinese 运动过程中的速度
  *
  * @discussion
- * [EN]: The speed during the sport activity, measured in meters per minute (m/min).
- * [CN]: 运动过程中的速度，以米每分钟表示（m/min）。
+ * [EN]: The speed during the sport activity, measured in meters per hour (m/h).
+ * [CN]: 运动过程中的速度，以米每小时表示（m/h）。
  *
  * @note
  * [EN]: Applicable to running, cycling, and similar activities.
  * [CN]: 适用于跑步、骑行等类似运动。
  */
-@property (nonatomic, assign) NSInteger speed;
+@property (nonatomic, assign) double speed;
 
 #pragma mark - Swimming Metrics
 /**
@@ -315,8 +315,8 @@ NS_ASSUME_NONNULL_BEGIN
  * @chinese 由 `TSSportDetailItemTable`（或同字段）查询结果数组构造明细模型列表
  *
  * @param dictionaryArray
- * EN: Array of per-sample/per-segment dictionaries (`userID`, `macAddress`, `sportID`, `type`, time fields, pace as s/km, speed as m/min, swim/jump/elliptical/rowing fields, etc.).
- * CN: 每条为一段明细的字典数组（含用户、设备、运动 ID、类型、时间戳、`pace` 秒/公里、`speed` 米/分钟及游泳/跳绳等扩展字段）。
+ * EN: Array of per-sample/per-segment dictionaries (`userID`, `macAddress`, `sportID`, `type`, time fields, pace as s/km, speed as m/h, swim/jump/elliptical/rowing fields, etc.).
+ * CN: 每条为一段明细的字典数组（含用户、设备、运动 ID、类型、时间戳、`pace` 秒/公里、`speed` 米/小时及游泳/跳绳等扩展字段）。
  *
  * @return
  * EN: Array of `TSSportItemModel`; empty array if input is nil or empty. Invalid entries are skipped.
